# backend/RestApiSwagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger PSWTM - OpenAPI 3.0",
    "description": "REST API Methods description for our Project",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:8800/api/v1"
    }
  ],
...
  "paths": {
...
    "/products": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "POST products as admin (Produkte Verwalten)",
        "description": "This can only be done by the logged in admin user.",
        "operationId": "createProducts",
        "requestBody": {
          "description": "Created product object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "coffee"
                  },
                  "costs": {
                    "type": "integer",
                    "format": "float",
                    "example": 2.6
                  }
                },
                "xml": {
                  "name": "product"
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product_id": {
                      "type": "integer",
                      "format": "int64",
                      "example": 10
                    },
                    "name": {
                      "type": "string",
                      "example": "coffee"
                    },
                    "costs": {
                      "type": "integer",
                      "format": "float",
                      "example": 2.6
                    }
                  },
                  "xml": {
                    "name": "product"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "products"
        ],
        "summary": "GET products by product name (Produkte Verwalten)",
        "description": "Can only used by admin",
        "operationId": "getProductByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The name that needs to be fetched. Use product1 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product_id": {
                      "type": "integer",
                      "format": "int64",
                      "example": 10
                    },
                    "name": {
                      "type": "string",
                      "example": "coffee"
                    },
                    "costs": {
                      "type": "integer",
                      "format": "float",
                      "example": 2.6
                    }
                  },
                  "xml": {
                    "name": "product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid productname supplied"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/products/list": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "GET all products (Produkte Verwalten)",
        "description": "Can be used by everbody",
        "operationId": "getProductsList",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The name that needs to be fetched. Use product1 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product_id": {
                      "type": "integer",
                      "format": "int64",
                      "example": 10
                    },
                    "name": {
                      "type": "string",
                      "example": "coffee"
                    },
                    "costs": {
                      "type": "integer",
                      "format": "float",
                      "example": 2.6
                    }
                  },
                  "xml": {
                    "name": "product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid productname supplied"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/products/{productID}": {
      "put": {
        "tags": [
          "products"
        ],
        "summary": "PUT existing product (Produkte Verwalten)",
        "description": "This can only be done by the logged in user admin.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "productname",
            "in": "path",
            "description": "name that need to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent product in the db",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 10
                  },
                  "product_name": {
                    "type": "string",
                    "example": "coffee"
                  },
                  "product_quantity": {
                    "type": "integer",
                    "example": 2
                  },
                  "product_costs": {
                    "type": "integer",
                    "format": "float",
                    "example": 2.6
                  }
                },
                "xml": {
                  "name": "product"
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "DELETE existing product (Produkte Verwalten)",
        "description": "This can only be done by the logged in user admin.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid product supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}